refactor : 유저 entity & MyProfile 엔티티변경
- 명확하게 역할을 구분하기 위해 기존 User, MyProfile을 Certification과 User로 변경함.

인증&개인정보 관련 정보는Certification에 저장하고 실제 서비스에 필요한 정보는 User테이블에 저장함.
